# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SewaSync - Gurdwara Management Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Expense {
  id: number;
  eventId: number;
  eventTitle?: string;
  description: string;
  amount: number;
  date: string;
  loggedBy?: string;
  createdAt: string;
}

export type EventType = (typeof EventType)[keyof typeof EventType];

export const EventType = {
  major_gurpurab: "major_gurpurab",
  regular_diwan: "regular_diwan",
  kirtan_darbar: "kirtan_darbar",
  amrit_sanchar: "amrit_sanchar",
  community_camp: "community_camp",
} as const;

export interface Event {
  id: number;
  title: string;
  date: string;
  type: EventType;
  volunteersNeeded: number;
  estimatedBudget: number;
  description?: string;
  createdAt: string;
}

export interface DashboardSummary {
  totalEvents: number;
  upcomingEvents: number;
  totalVolunteers: number;
  sewaFulfillmentPercent: number;
  totalEstimatedBudget: number;
  totalActualSpend: number;
  recentExpenses: Expense[];
  upcomingEventsList: Event[];
}

export type CreateEventBodyType =
  (typeof CreateEventBodyType)[keyof typeof CreateEventBodyType];

export const CreateEventBodyType = {
  major_gurpurab: "major_gurpurab",
  regular_diwan: "regular_diwan",
  kirtan_darbar: "kirtan_darbar",
  amrit_sanchar: "amrit_sanchar",
  community_camp: "community_camp",
} as const;

export interface CreateEventBody {
  title: string;
  date: string;
  type: CreateEventBodyType;
  volunteersNeeded: number;
  estimatedBudget: number;
  description?: string;
}

export interface BudgetStatus {
  eventId: number;
  eventTitle: string;
  estimatedBudget: number;
  actualSpend: number;
  variance: number;
  isOverBudget: boolean;
  expenses: Expense[];
}

export type VolunteerDepartment =
  (typeof VolunteerDepartment)[keyof typeof VolunteerDepartment];

export const VolunteerDepartment = {
  langar: "langar",
  joda_ghar: "joda_ghar",
  parking: "parking",
  kirtan_stage: "kirtan_stage",
  cleaning: "cleaning",
  security: "security",
} as const;

export interface Volunteer {
  id: number;
  eventId: number;
  eventTitle?: string;
  name: string;
  phone: string;
  department: VolunteerDepartment;
  registeredAt: string;
}

export type CreateVolunteerBodyDepartment =
  (typeof CreateVolunteerBodyDepartment)[keyof typeof CreateVolunteerBodyDepartment];

export const CreateVolunteerBodyDepartment = {
  langar: "langar",
  joda_ghar: "joda_ghar",
  parking: "parking",
  kirtan_stage: "kirtan_stage",
  cleaning: "cleaning",
  security: "security",
} as const;

export interface CreateVolunteerBody {
  eventId: number;
  name: string;
  phone: string;
  department: CreateVolunteerBodyDepartment;
}

export interface VolunteerFulfillment {
  eventId: number;
  eventTitle: string;
  eventDate: string;
  volunteersNeeded: number;
  volunteersRegistered: number;
  fulfillmentPercent: number;
}

export interface CreateExpenseBody {
  eventId: number;
  description: string;
  amount: number;
  date: string;
}

export type NotificationAudience =
  (typeof NotificationAudience)[keyof typeof NotificationAudience];

export const NotificationAudience = {
  all_sangat: "all_sangat",
  active_volunteers: "active_volunteers",
  event_volunteers: "event_volunteers",
} as const;

export type NotificationChannelsItem =
  (typeof NotificationChannelsItem)[keyof typeof NotificationChannelsItem];

export const NotificationChannelsItem = {
  sms: "sms",
  email: "email",
} as const;

export interface Notification {
  id: number;
  title: string;
  message: string;
  audience: NotificationAudience;
  channels: NotificationChannelsItem[];
  eventId?: number;
  sentAt: string;
  sentBy?: string;
}

export type SendNotificationBodyAudience =
  (typeof SendNotificationBodyAudience)[keyof typeof SendNotificationBodyAudience];

export const SendNotificationBodyAudience = {
  all_sangat: "all_sangat",
  active_volunteers: "active_volunteers",
  event_volunteers: "event_volunteers",
} as const;

export type SendNotificationBodyChannelsItem =
  (typeof SendNotificationBodyChannelsItem)[keyof typeof SendNotificationBodyChannelsItem];

export const SendNotificationBodyChannelsItem = {
  sms: "sms",
  email: "email",
} as const;

export interface SendNotificationBody {
  title: string;
  message: string;
  audience: SendNotificationBodyAudience;
  channels: SendNotificationBodyChannelsItem[];
  eventId?: number;
}

export interface ActivityLog {
  id: number;
  userId?: number;
  userName?: string;
  action: string;
  timestamp: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  super_admin: "super_admin",
  event_manager: "event_manager",
  sewadar: "sewadar",
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  lastLogin?: string;
  createdAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  super_admin: "super_admin",
  event_manager: "event_manager",
  sewadar: "sewadar",
} as const;

export interface CreateUserBody {
  username: string;
  name: string;
  role: CreateUserBodyRole;
  password: string;
}

export type ListEventsParams = {
  type?: string;
  upcoming?: boolean;
};

export type ListVolunteersParams = {
  eventId?: number;
  department?: string;
};

export type ListExpensesParams = {
  eventId?: number;
};