# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SewaSync - Gurdwara Management Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  userId?: number;
  userName?: string;
  action: string;
  timestamp: string;
}