# lib/api-zod/src/generated/types/createExpenseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SewaSync - Gurdwara Management Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseBody {
  eventId: number;
  description: string;
  amount: number;
  date: string;
}