# lib/api-zod/src/generated/types/createVolunteerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SewaSync - Gurdwara Management Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateVolunteerBodyDepartment } from "./createVolunteerBodyDepartment";

export interface CreateVolunteerBody {
  eventId: number;
  name: string;
  phone: string;
  department: CreateVolunteerBodyDepartment;
}