# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SewaSync - Gurdwara Management Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from "./event";
import type { Expense } from "./expense";

export interface DashboardSummary {
  totalEvents: number;
  upcomingEvents: number;
  totalVolunteers: number;
  sewaFulfillmentPercent: number;
  totalEstimatedBudget: number;
  totalActualSpend: number;
  recentExpenses: Expense[];
  upcomingEventsList: Event[];
}