# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SewaSync - Gurdwara Management Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  eventId: number;
  eventTitle?: string;
  description: string;
  amount: number;
  date: string;
  loggedBy?: string;
  createdAt: string;
}